+++ src/utility.jl
@@ -157,7 +157,7 @@ function _violated_eigen_vector(W::Array{<:Number}; tol = 1E-6)
     end
     return nothing
 end
-
+  
 """
     cheeger_constant(G::Matrix{<:Number}, 
                      optimizer::MOI.OptimizerWithAttributes; 
@@ -183,66 +183,82 @@ function cheeger_constant(
     G::Matrix{<:Number},
     optimizer::MOI.OptimizerWithAttributes;
     cheeger_ub = 1E4,
+    formulation_type = "set_cardinality", #set_cardinality, set_volume
     optimizer_log = false,
 )
-    if size(G)[1] !== size(G)[2]
+    num_nodes = size(G)[1]
+    if num_nodes != size(G)[2]
         Memento.error(_LOGGER, "Input adjacency has to be a square matrix")
     end
-    num_nodes = size(G)[1]
 
     result_cheeger = Dict{String,Any}("cheeger_constant" => 0)
 
-    if !isapprox(LOpt.algebraic_connectivity(G), 0, atol = 1E-5)
-        m_cheeger = JuMP.Model(optimizer)
-        if !optimizer_log
-            JuMP.set_silent(m_cheeger)
-        end
+    (isapprox(LOpt.algebraic_connectivity(G), 0, atol = 1E-5)) && result_cheeger
 
-        # Variables
-        JuMP.@variable(m_cheeger, 0 <= z[1:num_nodes] <= 1, Bin)
-        JuMP.@variable(m_cheeger, -1 <= Z[1:num_nodes, 1:num_nodes] <= 1)
-        JuMP.@variable(m_cheeger, 0 <= cheeger <= cheeger_ub)
-        JuMP.@variable(m_cheeger, cheeger_z[1:num_nodes] >= 0)
+    m_cheeger = JuMP.Model(optimizer)
+    (!optimizer_log) && (JuMP.set_silent(m_cheeger))
 
-        # Constraints
+    JuMP.@variable(m_cheeger, 0 <= z[1:num_nodes] <= 1, Bin)
+    JuMP.@variable(m_cheeger, 0 <= Z[1:num_nodes, 1:num_nodes] <= 1)
+    JuMP.@variable(m_cheeger, 0 <= cheeger <= cheeger_ub)
+    JuMP.@variable(m_cheeger, cheeger_z[1:num_nodes] >= 0)
+
+    if formulation_type == "set_cardinality"
         JuMP.@constraint(m_cheeger, sum(z) >= 1)
         JuMP.@constraint(m_cheeger, sum(z) <= floor(num_nodes / 2))
+
         JuMP.@constraint(
             m_cheeger,
             sum(cheeger_z) >= sum(
-                G[i, j] * Z[i, j] for i in 1:(num_nodes-1), j in (i+1):num_nodes if
+                G[i, j] * (z[i] + z[j] - 2 * Z[i, j]) for
+                i in 1:(num_nodes-1), j in (i+1):num_nodes if
                 !(isapprox(G[i, j], 0, atol = 1E-5))
             )
         )
 
+    elseif formulation_type == "set_volume"
+        laplacian = LOpt.laplacian_matrix(float(G .> 0))
+        k_vol_ub = sum(G .> 0) / 2  #|E|
+        JuMP.@constraint(m_cheeger, sum(z[i] * laplacian[i, i] for i in 1:num_nodes) >= 1)
         JuMP.@constraint(
             m_cheeger,
-            [i = 1:(num_nodes-1), j = (i+1):num_nodes],
-            Z[i, j] >= z[i] - z[j]
+            sum(z[i] * laplacian[i, i] for i in 1:num_nodes) <= k_vol_ub
         )
+
         JuMP.@constraint(
             m_cheeger,
-            [i = 1:(num_nodes-1), j = (i+1):num_nodes],
-            Z[i, j] >= z[j] - z[i]
+            sum(cheeger_z[i] * laplacian[i, i] for i in 1:num_nodes) >= sum(
+                G[i, j] * (z[i] + z[j] - 2 * Z[i, j]) for
+                i in 1:(num_nodes-1), j in (i+1):num_nodes if
+                !(isapprox(G[i, j], 0, atol = 1E-5))
+            )
         )
-
-        for i in 1:num_nodes
-            LOpt.relaxation_bilinear(m_cheeger, cheeger_z[i], cheeger, z[i])
-        end
-
-        # Objective
-        JuMP.@objective(m_cheeger, Min, cheeger)
-
-        JuMP.optimize!(m_cheeger)
-
-        result_cheeger["cheeger_constant"] = JuMP.objective_value(m_cheeger)
-        result_cheeger["S"] =
-            Set(findall(isapprox.(abs.(JuMP.value.(z)), 1, atol = 1E-5)))
-        result_cheeger["S_complement"] =
-            Set(findall(isapprox.(abs.(JuMP.value.(z)), 0, atol = 1E-5)))
-        result_cheeger["solve_time"] = JuMP.solve_time(m_cheeger)
     end
 
+    for i in 1:(num_nodes-1), j in (i+1):num_nodes
+        if !(isapprox(G[i, j], 0, atol = 1E-5))
+            LOpt.relaxation_bilinear(m_cheeger, Z[i, j], z[i], z[j])
+        end
+    end
+
+    for i in 1:num_nodes
+        LOpt.relaxation_bilinear(m_cheeger, cheeger_z[i], cheeger, z[i])
+    end
+
+    # Objective
+    JuMP.@objective(m_cheeger, Min, cheeger)
+    JuMP.optimize!(m_cheeger)
+
+    result_cheeger["cheeger_constant"] = JuMP.objective_value(m_cheeger)
+    result_cheeger["S"] = Set(findall(isapprox.(abs.(JuMP.value.(z)), 1, atol = 1E-5)))
+    result_cheeger["S_complement"] =
+        Set(findall(isapprox.(abs.(JuMP.value.(z)), 0, atol = 1E-5)))
+    result_cheeger["solve_time"] = JuMP.solve_time(m_cheeger)
+    result_cheeger["termination_status"] = JuMP.termination_status(m_cheeger)
+    result_cheeger["optimality_gap_percent"] = 100 * JuMP.relative_gap(m_cheeger)
+
+    (length(result_cheeger["S"]) == 0) &&
+        (Memento.error(_LOGGER, "Cheeger set cardinality cannot be zero."))
     return result_cheeger
 end
 
@@ -325,7 +341,7 @@ function _PMinorIdx(
 
     return minor_idx_dict
 end
-
+                    
 function _vertices_with_augment_edges(data::Dict{String,Any})
     v = Set{Int64}()
     adj = data["adjacency_augment_graph"]
@@ -355,7 +371,7 @@ end
     weighted_adjacency_matrix(G::Graphs.SimpleGraphs.SimpleGraph{<:Number}, adjacency_augment_graph::Array{<:Number}, size::Int64)
 
 Returns a weighted adjacency matrix for 
-the connected part of  graph.  
+the connected part of graph.  
 """
 function weighted_adjacency_matrix(
     G::Graphs.SimpleGraphs.SimpleGraph{<:Number},
@@ -411,7 +427,6 @@ end
 
 Returns all combinations possible for given `n` and `k`.
 """
-
 function edge_combinations(num_edges::Int64, kopt_parameter::Int64)
     if kopt_parameter < 2 || kopt_parameter > 3
         Memento.error(_LOGGER, "kopt_parameter must be either 2 or 3")
@@ -428,7 +443,7 @@ function edge_combinations(num_edges::Int64, kopt_parameter::Int64)
         ]
     end
 end
-
+                                          
 """
     add_multiple_edges!(G::Graphs.SimpleGraphs.SimpleGraph{<:Number},edge_set::Vector{Tuple{Int64,Int64}},)